Move example objects out of GeometricSimplicialComplex.scala
