+++ src/main/scala/DimensionReduction/GeometricSimplicialComplex.scala
@@ -397,152 +397,3 @@ object GeometricSimplicialComplex {
   }
 
 }
-
-// This set of tests currently passes. (Be warned - it takes a long time to run...)
-object ProjectionTest extends App {
-  val maybeCulledGeomComp: Option[GeometricSimplicialComplex] = GeometricSimplicialComplex.loadFromFile("Datasets/CulledNormalizedIrisComplex.txt")
-
-  assert(maybeCulledGeomComp.nonEmpty)
-
-  val culledGeomComp: GeometricSimplicialComplex = maybeCulledGeomComp.get
-
-  val numPointsToTest: Int = 100
-  val uniformHypercubePoints: Vector[Point] = (0 until numPointsToTest).map(_ => Point((0 until 4).map(_ => scala.util.Random.nextDouble()).toVector)).toVector
-
-  uniformHypercubePoints.zipWithIndex.foreach({case (point, idx) =>
-
-    println("Testing point " + (idx+1) + " of " + numPointsToTest)
-    println(point)
-
-    val projection: Point = culledGeomComp.project(point)
-    val bruteForceProjection: Point = culledGeomComp.bruteForceProject(point)
-    val projectionOfProjection: Point = culledGeomComp.project(projection)
-
-    // A projection operator should be idempotent.
-    assert(projection == projectionOfProjection)
-
-    // Our intelligent operator should produce correct results
-    assert(projection == bruteForceProjection)
-
-    // Our complex should actually contain the projections of points onto the complex.
-    assert(culledGeomComp.simplices.exists(_.containsPoint(projection)))
-  })
-
-}
-
-object LoadTest extends App {
-
-  println("Loading...")
-
-  val loadedGeomComp: Option[GeometricSimplicialComplex] = GeometricSimplicialComplex.loadFromFile("Datasets/UnculledNormalizedIrisComplex.txt")
-
-  println("Loaded!")
-
-  println(loadedGeomComp)
-
-
-  loadedGeomComp match {
-    case Some(complex) =>
-      complex.printStatistics(30)
-    case None =>
-      println("Red alert! Failed to load simplicial complex from file!")
-  }
-
-
-
-  val culledGeomComp: Option[GeometricSimplicialComplex] = GeometricSimplicialComplex.loadFromFile("Datasets/CulledNormalizedIrisComplex.txt")
-
-  culledGeomComp match {
-    case Some(complex) =>
-      complex.printStatistics(8)
-      println("Brute force projection...")
-      println(complex.project(Point(Vector(0.0, 0.0, 0.0, 0.0))))
-      println(complex.project(Point(Vector(1.0, 1.0, 1.0, 1.0))))
-      println(complex.project(Point(Vector(0.9444444444444444, 0.7499999999999998, 0.9661016949152542, 0.8750000000000001))))
-
-      // Choose a single vertex from the complex...
-      println(complex.project(complex.simplices.head.points.head))
-    case None =>
-      println("Red alert! Failed to load simplicial complex from file!")
-  }
-
-}
-
-
-
-object GeometricSimplicialComplexTest extends App {
-
-  val points: Vector[Point] = Vector(
-    Point(Vector(0.0, 0.0, 0.0)),
-    Point(Vector(1.0, 0.0, 0.0)),
-    Point(Vector(0.0, 1.0, 0.0)),
-    Point(Vector(0.0, 0.0, 1.0)),
-    Point(Vector(1.0, 1.0, 1.0))
-  )
-
-  val pointNames: Point => String = {
-    case Point(Vector(0.0, 0.0, 0.0)) => "O"
-    case Point(Vector(1.0, 0.0, 0.0)) => "X"
-    case Point(Vector(0.0, 1.0, 0.0)) => "Y"
-    case Point(Vector(0.0, 0.0, 1.0)) => "Z"
-    case Point(Vector(1.0, 1.0, 1.0)) => "S"
-    case _ => "Unknown Point"
-  }
-
-  val delaunay: Vector[PointedAffineSpace] = Delaunay.DelaunayUtil.getDelaunaySimplicialization(points)
-
-  delaunay.foreach((space: PointedAffineSpace) => {
-    println("(" + space.vertices.map(pointNames).reduce(_ + ", " + _) + ")")
-  })
-
-  val simplicialization: GeometricSimplicialComplex = new GeometricSimplicialComplex(delaunay)
-
-  println(simplicialization)
-
-  simplicialization.nSimplices foreach { case (n, nSimplices) =>
-    println(n + "-Simplices:")
-    nSimplices.foreach((simplex: Simplex) =>
-      println(if (simplex.points.nonEmpty) {"[" + simplex.points.map(pointNames).reduce(_ + ", " + _) + "]"} else "[]")
-    )
-  }
-
-}
-
-
-object GeometricSimplicialComplexTest2 extends App {
-
-  val points: Vector[Point] = Vector(
-    Point(Vector(0.0, 0.0)),
-    Point(Vector(3.0, 4.0)),
-    Point(Vector(4.0, 3.0)),
-    Point(Vector(0.0, 25.0/7.0)),
-  )
-
-  val k = 25.0 / 7.0
-
-  val pointNames: Point => String = {
-    case Point(Vector(0.0, 0.0)) => "O"
-    case Point(Vector(3.0, 4.0)) => "Y"
-    case Point(Vector(4.0, 3.0)) => "X"
-    case Point(Vector(0.0, k)) => "S"
-    case _ => "Unknown Point"
-  }
-
-  val delaunay: Vector[PointedAffineSpace] = Delaunay.DelaunayUtil.getDelaunaySimplicialization(points)
-
-  delaunay.foreach((space: PointedAffineSpace) => {
-    println("(" + space.vertices.map(pointNames).reduce(_ + ", " + _) + ")")
-  })
-
-  val simplicialization: GeometricSimplicialComplex = new GeometricSimplicialComplex(delaunay)
-
-  println(simplicialization)
-
-  simplicialization.nSimplices foreach { case (n, nSimplices) =>
-    println(n + "-Simplices:")
-    nSimplices.foreach((simplex: Simplex) =>
-      println(if (simplex.points.nonEmpty) {"[" + simplex.points.map(pointNames).reduce(_ + ", " + _) + "]"} else "[]")
-    )
-  }
-
-}
